requirements.txt
python-telegram-bot==13.13
prettytable==3.3.0